requirements.txt
fastapi>=0.68.0,<0.69.0
pydantic>=1.8.0,<2.0.0
uvicorn>=0.15.0,<0.16.0
python-dotenv>=0.19.2
authlib>=1.0
requests>=2.27.1
itsdangerous>=1.1.0
httpx>=0.24.0